test_cpp_st_numbering.py: Added connectivity check to V1 and V2 subgraphs. make-st-numbering.cpp: tried replacing "double" with "long double" to eliminate 		       floating point precision problems. Didn't work.

+++ wykobi-master/make-st-numbering.cpp
@@ -13,7 +13,7 @@
 using namespace std;
 namespace pt = boost::property_tree;
 
-typedef double ord_type;
+typedef long double ord_type;
 typedef vector<ord_type> position_vec;
 typedef pair<ord_type, ord_type> coordinate;
 typedef vector<coordinate> coordinate_vec;
